* Fixes for PHP 8.x compatibility * Implement changing poster image for video files * Fixed how additional buttons are displayed on attachment detail pages * Fixed downloading video with video player * Fix for getting rendition URL for mux encoded videos * Updated video.js library to 7.20.3 * Updated related video.js plugins to latest versions * Fix mux transfer task to work with large files and give better feedback.  Note: Mux transfers wait until the mp4 renditions are encoded by mux before transferring.  This can take a while for large files. * Remove preflight admin notice * Added support for Supabase Storage.  This support is only provided for specific development scenarios, you should not use Supabase Storage for general purpose storage because you think it's inexpensive.  Supabase Storage support is for when you need to *integrate* with an application already using Supabase. * Added a new `Transfer Mux Video` task that allows you to transfer encoded videos from Mux to cloud or local storage to serve the videos from other services and CDNs instead of Mux.  Use Mux to encode, serve the result elsewhere. * Added a `Regenerate Filmstrip` task to generate new filmstrips for Mux encoded videos that might not have had them generated. * The Gutenberg block now supports outputting the video's dimension to insure that there isn't any reflow on document load. * Fixed INFO log messages with regards to PHP 8.1. * Transferred Mux videos display HLS logo instead of the Mux logo in the media library. * In the list view in the Media Library, the Mux or HLS logo is now displayed. * The `Relink Mux Video` task will relink any videos that were transferred with the `Transfer Mux Video` task unless they've been deleted on Mux. * You can setup Media Cloud to automatically transfer Mux encoded videos to local or cloud storage after they've finished being encoded by Mux. * Added `wp mediacloud:video transfer` command to transfer videos from Mux to cloud or local storage. * Added `wp mediacloud:video relink` command to relink videos that were transferred from Mux to local or cloud storage. * Added toggle to `Video Player` settings to replace the WordPress admin default video player with the native video player using HLS.js.  This allows you to preview Mux encoded videos in the admin without having to create a post. * Added toggle to override the default WordPress video shortcode and replace it with the Media Cloud Video Player.  You should, however, consider switching to the Media Cloud Video Player shortcode instead because it performs better. * The Elementor Media Cloud Video Player widget now supports playing non-Mux videos. * Added `hls` object to WP REST Attachment object that includes the playlist url, poster image url, filmstrip url, width, height, duration, mp4 fallback rendition url and subtitles. * Added the ability to add captions (via a VTT text file) to any video, regardless if it's Mux encoded or not.  This feature is available on the edit detail screen for a video attachment.  You must be using the Media Cloud Video Player feature. * Removed `yuloh/bccomp-polyfill` dependency.

+++ classes/Tools/Video/Driver/Mux/MuxHooks.php
@@ -25,6 +25,7 @@ use  MediaCloud\Vendor\MuxPhp\Models\InputSettings ;
 use  MediaCloud\Vendor\MuxPhp\Models\PlaybackPolicy ;
 use function  MediaCloud\Plugin\Utilities\arrayPath ;
 use function  MediaCloud\Plugin\Utilities\gen_uuid ;
+use function  MediaCloud\Plugin\Utilities\postIdExists ;
 class MuxHooks
 {
     /** @var MuxToolSettings|MuxToolProSettings  */
@@ -59,6 +60,9 @@ class MuxHooks
             );
         }
         
+        if ( is_admin() ) {
+            add_action( 'wp_ajax_mcloud_replace_poster', [ $this, 'ajaxReplacePoster' ] );
+        }
         add_filter( 'template_include', [ $this, 'handleWebhook' ] );
     }
     
@@ -255,7 +259,7 @@ class MuxHooks
                 __METHOD__,
                 __LINE__
             );
-            $this->generateFilmstripForAttachment( $asset );
+            $asset->generateFilmstrip();
             return;
         }
         
@@ -312,28 +316,7 @@ class MuxHooks
         update_post_meta( $thumbId, '_wp_attached_file', $thumbAttachmentMeta['file'] );
         update_post_meta( $asset->attachmentId, '_thumbnail_id', $thumbId );
         wp_update_attachment_metadata( $thumbId, $thumbAttachmentMeta );
-        $this->generateFilmstripForAttachment( $asset );
-    }
-    
-    /**
-     * @param MuxAsset $asset
-     *
-     * @throws \Freemius_Exception
-     */
-    protected function generateFilmstripForAttachment( $asset )
-    {
-        Logger::info(
-            'Mux: generateFilmstripForAttachment',
-            [],
-            __METHOD__,
-            __LINE__
-        );
-        Logger::warning(
-            'Mux: generateFilmstripForAttachment could not be run, not premium',
-            [],
-            __METHOD__,
-            __LINE__
-        );
+        $asset->generateFilmstrip();
     }
     
     public function handleStaticRenditionsReady( $jsonData )
@@ -384,6 +367,7 @@ class MuxHooks
             }
         
         }
+        $asset->backgroundTransfer();
     }
     
     public function handleAssetReady( $jsonData )
@@ -496,6 +480,13 @@ class MuxHooks
         if ( empty($asset) ) {
             return;
         }
+        
+        if ( $asset->isTransferred == 1 ) {
+            $asset->isDeleted = 1;
+            $asset->save();
+            return;
+        }
+        
         $asset->delete();
     }
     
@@ -710,6 +701,29 @@ class MuxHooks
     public function handleUpload( $attachmentId, $file, $meta )
     {
         $this->handleDirectUpload( $attachmentId, $meta );
+    }
+    
+    public function ajaxReplacePoster()
+    {
+        wp_verify_nonce( $_POST['nonce'], 'media-cloud-mux-replace-poster' );
+        $attachmentId = arrayPath( $_POST, 'attachmentId', null );
+        if ( empty($attachmentId) ) {
+            wp_send_json_error( 'Missing attachment ID', 400 );
+        }
+        $newPosterId = arrayPath( $_POST, 'posterId', null );
+        if ( empty($newPosterId) ) {
+            wp_send_json_error( 'Missing poster ID', 400 );
+        }
+        if ( !postIdExists( $attachmentId ) ) {
+            wp_send_json_error( 'Invalid attachment ID', 400 );
+        }
+        if ( !postIdExists( $newPosterId ) ) {
+            wp_send_json_error( 'Invalid poster ID', 400 );
+        }
+        update_post_meta( $attachmentId, '_thumbnail_id', $newPosterId );
+        wp_send_json( [
+            'success' => true,
+        ] );
     }
 
 }
